+++ camat/__init__.py
@@ -1,7 +1,8 @@
-__version__ = "0.1.5"
+__version__ = "0.1.6"
 
 from .pattern_search import run_pattern_search
 from .parser_registry import get_parse_files, list_parsers, parse_files_quiet
+from .mensural_backend import parse_files_mensural
 from .verovio_render import (
     get_toolkit as vrv_get_toolkit,
     vrv_set_options,
@@ -75,6 +76,7 @@ __all__ = [
     'get_parse_files',
     'list_parsers',
     'parse_files_quiet',
+    'parse_files_mensural',
     'vrv_get_toolkit',
     'vrv_set_options',
     'vrv_guess_input_from',